requirements.txt
blinker==1.4
cffi==1.8.3
click==6.6
cryptography==1.5.2
cssmin==0.2.0
decorator==4.0.10
Flask==0.11.1
Flask-Assets==0.12
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.0
Flask-Injector==0.7.1
Flask-Login==0.3.2
Flask-OAuthlib==0.9.3
Flask-PyMongo==0.4.1
Flask-Restless==0.17.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
httplib2==0.9.2
idna==2.1
injector==0.10.1
itsdangerous==0.24
Jinja2==2.8
jsmin==2.2.1
MarkupSafe==0.23
mimerender==0.6.0
oauth2==1.9.0.post1
oauthlib==2.0.0
paramiko==2.0.2
pbr==1.10.0
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.14
pymongo==3.3.0
pysftp==0.2.9
python-dateutil==2.5.3
python-mimeparse==1.5.5
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
SQLAlchemy==1.1.0
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
webassets==0.12.0
Werkzeug==0.11.11
WTForms==2.1
